MARM-41 1) Рефакторинг: перенес архивацию из класса LocalConfStorage в ArchiveService 2) Изменил имена контейнеров и образов докера + добавил подключение к сетям cms и db 3) Добавил проверку буффера при отправке данных в БД

+++ src/models/RemoteConfStorage.php
@@ -98,6 +98,10 @@ class RemoteConfStorage {
      * @param string $collectionName
      */
     public function flush(string $collectionName) {
+        if (empty($this->buffer)) {
+            return;
+        }
+
         $collection = $this->getDb()->selectCollection($collectionName);
         $collection->deleteMany([
             '_id' => [
